requirements.txt
-e .
aws-cdk-lib==2.12.0
constructs>=10.0.0,<11.0.0
aws-cdk.core
aws-cdk.aws_s3attrs==21.4.0
aws-cdk.aws-events==1.144.0
aws-cdk.aws-iam==1.144.0
aws-cdk.aws-kms==1.144.0
aws-cdk.aws-s3==1.144.0
aws-cdk.cloud-assembly-schema==1.144.0
aws-cdk.core==1.144.0
aws-cdk.cx-api==1.144.0
aws-cdk.region-info==1.144.0
cattrs==1.10.0
constructs==3.3.212
jsii==1.52.1
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.0.1
attrs==21.4.0
aws-cdk.aws-events==1.144.0
aws-cdk.aws-iam==1.144.0
aws-cdk.aws-kms==1.144.0
aws-cdk.aws-s3==1.144.0
aws-cdk.cloud-assembly-schema==1.144.0
aws-cdk.core==1.144.0
aws-cdk.cx-api==1.144.0
aws-cdk.region-info==1.144.0
cattrs==1.10.0
constructs==3.3.212
jsii==1.52.1
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.0.1
